play from url

+++ house_bot.py
@@ -13,6 +13,8 @@ import asyncio
 import json
 import time
 import random
+import os
+import youtube_dl
 
 # read config.json to get token and botid
 config = open('config.json', 'r')
@@ -54,29 +56,87 @@ async def respond(message):
             time.sleep(round(len(str(response)) / 20, 0))
         await message.channel.send(response)
 
-@bot.command(name='gethappy')
-async def gethappy(ctx):
-    # Gets voice channel of message author
+"""
+Music Bot Functionality
+"""
+servers = {}
+
+async def getYoutubeVideoInfo(url):
+    ydl_opts = {
+        'format': 'bestaudio/best',
+        'outtmpl': 'downloadedsongs/%(title)s.%(ext)s',
+        'postprocessors': [{
+            'key': 'FFmpegExtractAudio',
+            'preferredcodec': 'mp3',
+            'preferredquality': '192',
+        }],
+    }
+    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
+        song_info = ydl.extract_info(url, download=False)
+    # actual song url is song_info["formats"][0]["url"]
+    return song_info
+
+async def downloadYoutubeVideo(url, ctx):
+    outputPath = 'songs/' + str(ctx.guild.id) + '/' + '%(title)s.%(ext)s'
+    ydl_opts = {
+        'format': 'bestaudio/best',
+        'outtmpl': outputPath,
+        'postprocessors': [{
+            'key': 'FFmpegExtractAudio',
+            'preferredcodec': 'mp3',
+            'preferredquality': '192',
+        }],
+    }
+    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
+        video = ydl.download([url])
+
+@bot.command(name='play')
+async def play(ctx):
+    if ctx.author.voice == None:
+        await ctx.send(str(ctx.author.name) + " is not in a channel.")
+        return
+
     voice_channel = ctx.author.voice.channel
     channel = None
     if voice_channel != None:
         channel = voice_channel.name
-        vc = await voice_channel.connect()
-        vc.play(discord.FFmpegPCMAudio(source="gethappy.mp3")) # Sleep while audio is playing.
 
-        while vc.is_playing(): # prevents the bot from disconnecting while it is playing
-            #time.sleep(.1)
-            await asyncio.sleep(1)
+    if ctx.guild.id not in servers:
+        servers[ctx.guild.id] = {'queue': []}
 
-        await vc.disconnect()
-    else:
-        await ctx.send(str(ctx.author.name) + "is not in a channel.")
-    # Delete command after the audio is done playing.
+    server = servers[ctx.guild.id]
+
+    args = str(ctx.message.content).split(" ")
+    if len(args) > 0:
+        server['queue'].append(args[1])
+
+    print(server['queue'])
     await ctx.message.delete()
+
+    if ctx.voice_client != None:
+        return
+
+    vc = await voice_channel.connect()
+    url = server['queue'].pop()
+    video = await downloadYoutubeVideo(url, ctx)
+    video_info = await getYoutubeVideoInfo(url)
+    path = 'songs\\' + str(ctx.guild.id) + '\\' + video_info['title'] +'.mp3'
+    vc.play(discord.FFmpegPCMAudio(source=path))
+    # prevents the bot from disconnecting while it is playing
+    while vc.is_playing():
+        #time.sleep(.1)
+        await asyncio.sleep(1)
+    os.remove(path) # delete downloaded audio after playback
+
+    if server['queue'] != 0:
+        play(ctx)
+
+    await vc.disconnect()
 
 @bot.command(name="leave")
 async def leave(ctx):
-    if (ctx.voice_client.is_connected()):
+    if ctx.voice_client.is_connected() or ctx.voice_client.is_playing():
+        await ctx.voice_client.stop()
         await ctx.voice_client.disconnect()
 
 @bot.command(name="pause")
@@ -87,4 +147,17 @@ async def pause(ctx):
 async def leave(ctx):
     ctx.voice_client.resume()
 
+
+
+
 bot.run(token)
+
+"""
+@bot.command(pass_context=True, brief="This will play a song 'play [url]'", aliases=['pl'])
+async def play(ctx, url:str):
+    server = ctx.message.server
+    voice_client = bot.voice_client_in(server)
+    player = await voice_client.create_ytdl_player(url, after=lambda: check_queue(server.id), before_options='-reconnect 1 -reconnect_streamed 1 -reconnect_delay_max 5')
+    players[server.id] = player
+    player.start()
+"""
